--Added a sample entity and cache implementation to demonstrate behavior. --Minor refactor
