+++ src/main/java/com/rambler/inmem/entity/SampleEntity.java
@@ -1,0 +1,9 @@
+package com.rambler.inmem.entity;
+
+public class SampleEntity {
+
+    String id;
+
+    String value;
+
+}
