+++ src/main/java/com/rambler/inmem/repository/impl/GenericKeyValueRepository.java
@@ -1,20 +1,62 @@
 package com.rambler.inmem.repository.impl;
 
-import java.util.List;
+import com.github.benmanes.caffeine.cache.Cache;
+import com.github.benmanes.caffeine.cache.Caffeine;
+import org.springframework.beans.factory.annotation.Value;
+
+import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
+import java.util.concurrent.TimeUnit;
 
-/**
- * A generic contract for KeyValue type repository implementation
- *
- * Provides various behavior to operate upon a KeyValue store
- *
- * @param <Key> Key Type
- * @param <Val> Value Type
- *
- * @author vandit
- */
-public interface GenericKeyValueRepository<Key, Val> extends GenericRepository {
+public abstract class GenericKeyValueRepository<Key,Val> implements IGenericKeyValueRepository<Key,Val> {
+
+    @Value("${repo.cache.max.allowed.size:10000000}")
+    protected int maxSize;
+
+    @Value("${repo.cache.max.default.size:10000}")
+    protected int defaultSize;
+
+    @Value("${repo.cache.max.allowed.ttl:10000000}")
+    protected int maxTtl;
+
+    @Value("${repo.cache.max.default.ttl:10000000}")
+    protected int defaultTtl;
+
+    private Cache<Key,Val> cache;
+
+
+    /**
+     * Registry to hold all caches
+     */
+    private final Map<String, Cache<Key,Val>> cacheRegistry=new HashMap<>();
+
+
+    public GenericKeyValueRepository(String name){
+        init(name,defaultTtl,defaultSize);
+    }
+
+    public GenericKeyValueRepository(String name, int ttl){
+        init(name,ttl,defaultSize);
+    }
+
+    public GenericKeyValueRepository(String name, int ttl, int size){
+        init(name,ttl,size);
+    }
+
+    /**
+     * Initializes caffeine cache instance with provided attributes
+     * @param name : Name of the cache
+     * @param ttl : Time to live for a single entry in the cache
+     * @param size : Maximum allowed entries in the cache
+     */
+    private void init(String name,int ttl,int size){
+        cache=Caffeine.newBuilder()
+                .maximumSize(size>maxSize?maxSize:size)
+                .expireAfterAccess(ttl>maxTtl?maxTtl:ttl, TimeUnit.MILLISECONDS)
+                .build();
+    }
+
 
     /**
      * Fetches value for the provided key
@@ -22,7 +64,10 @@ public interface GenericKeyValueRepository<Key, Val> extends GenericRepository {
      * @param key : The lookup key
      * @return the value associated to the provided key. `null` if not found.
      */
-    Val get(Key key);
+    @Override
+    public Val get(Key key) {
+        return cache.getIfPresent(key);
+    }
 
     /**
      * Fetches value for the provided key
@@ -30,7 +75,10 @@ public interface GenericKeyValueRepository<Key, Val> extends GenericRepository {
      * @param keys : A set of keys to be looked up
      * @return a map key-value pairs.Associated value for the key will be `null` if not found.
      */
-    Map<Key, Val> get(Set<Key> keys);
+    @Override
+    public Map<Key, Val> get(Set<Key> keys) {
+        return cache.getAllPresent(keys);
+    }
 
     /**
      * Fetches all the values for the underlying entity
@@ -38,21 +86,30 @@ public interface GenericKeyValueRepository<Key, Val> extends GenericRepository {
      * @param size : A set of keys to be looked up
      * @return a map key-value pairs.Associated value for the key will be `null` if not found.
      */
-    Map<Key, Val> getAll(int size);
+    @Override
+    public Map<Key, Val> getAll(int size) {
+        throw new UnsupportedOperationException("This is not supported now");
+    }
 
     /**
      * Fetches all the values for the underlying entity
      *
      * @return a map key-value pairs.
      */
-    Map<Key, Val> getAll();
+    @Override
+    public Map<Key, Val> getAll() {
+        return cache.asMap();
+    }
 
     /**
      * Returns all the keys for the entity
      *
      * @return A set of all the keys in the repo
      */
-    Set<Key> keys();
+    @Override
+    public Set<Key> keys() {
+        return cache.asMap().keySet();
+    }
 
     /**
      * Returns all the keys for the entity , fetched in batches for better performance
@@ -60,34 +117,50 @@ public interface GenericKeyValueRepository<Key, Val> extends GenericRepository {
      * @param batchSize : size of a single batch
      * @return A set of all the keys in the repo
      */
-    Set<Key> keys(int batchSize);
+    @Override
+    public Set<Key> keys(int batchSize) {
+        throw new UnsupportedOperationException("This is not supported");
+    }
 
     /**
      * Adds a new entry in the store ,  overwrites existing if already exist
-     * @param key : Key
-     * @param val : Value
+     *
+     * @param key   : Key
+     * @param value : Value
      * @return : true denotes a successful operation , false denotes a failure
      */
-    boolean put(Key key, Val val);
-
+    @Override
+    public boolean put(Key key, Val value) {
+        cache.put(key,value);
+        return true;
+    }
 
     /**
      * Adds a new entry in the store only if it does not exist already
-     * @param key : Key
-     * @param val : Value
+     *
+     * @param key   : Key
+     * @param value : Value
      * @return : true denotes a successful put , false if key already exist
      */
-    boolean putIfAbsent(Key key, Val val);
-
+    @Override
+    public boolean putIfAbsent(Key key, Val value) {
+        if(cache.getIfPresent(key)==null){
+            return put(key,value);
+        }else
+            return false;
+    }
 
     /**
      * Adds a new entry in the store ,overwrites existing if already exist
-     * @param entries : A map of key values to be inserted into store
      *
+     * @param entries : A map of key values to be inserted into store
      * @return : true denotes a successful operation , false denotes a failure
      */
-    boolean put(Map<Key, Val> entries);
-
+    @Override
+    public boolean put(Map<Key, Val> entries) {
+        cache.putAll(entries);
+        return true;
+    }
 
     /**
      * Removes the key from the store
@@ -95,14 +168,21 @@ public interface GenericKeyValueRepository<Key, Val> extends GenericRepository {
      * @param key : key to be removed
      * @return true if successful , false if key does not exist
      */
-    boolean delete(Key key);
+    @Override
+    public boolean delete(Key key) {
+        cache.invalidate(key);
+        return true;
+    }
 
     /**
      * Removes the key from the store
      *
      * @param keys : keys to be removed
-     * @return true if successful , false if key does not exist
+     * @return List of booleans , each entry denoting : true if successful , false if key does not exist
      */
-    boolean delete(Set<Key> keys);
-
+    @Override
+    public boolean delete(Set<Key> keys) {
+        cache.invalidateAll(keys);
+        return true;
+    }
 }
