+++ src/main/java/com/rambler/inmem/repository/impl/IGenericRepository.java
@@ -5,5 +5,5 @@ package com.rambler.inmem.repository.impl;
  * A generic contract for a repository
  * @author rambler
  */
-public interface GenericRepository {
+public interface IGenericRepository {
 }
